Refactor main.rs by splitting into app module
Reduces main.rs from 1804 lines to 119 lines (93% reduction) by extracting
code into organized modules under src/app/.

Structure:
- app/mod.rs: App struct and core logic (225 lines)
- app/event_handler.rs: Event and key handling (654 lines)
- app/actions.rs: Worker operation methods (494 lines)
- app/rendering.rs: UI rendering methods (261 lines)
- app/types.rs: Type definitions (76 lines)
- app/worker_view.rs: WorkerView struct (43 lines)

This improves code maintainability and readability by separating concerns
into focused modules with clear responsibilities.
